domainmodel/src/test/java/de/staatsbibliothek/berlin/hsp/domainmodel/HSP_UUI_FactoryTest.java
/*
 * MIT License
 *
 * Copyright (c) 2023 Staatsbibliothek zu Berlin - Preußischer Kulturbesitz
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in all
 * copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 * SOFTWARE.
 */

package de.staatsbibliothek.berlin.hsp.domainmodel;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertNotNull;
import static org.junit.jupiter.api.Assertions.assertTrue;

import java.nio.charset.StandardCharsets;
import org.junit.jupiter.api.Test;

/**
 * @author konrad.eichstaedt@sbb.spk-berlin.de on 26.08.22.
 * @version 1.0
 */
public class HSP_UUI_FactoryTest {

  @Test
  void testGenerate() {
    assertNotNull(HSP_UUI_Factory.generate());
    assertTrue(HSP_UUI_Factory.generate().contains("HSP-"));
  }

  @Test
  void testGenerateFromByte() {
    assertNotNull(HSP_UUI_Factory.generate("3434".getBytes(StandardCharsets.UTF_8)));
    assertTrue(HSP_UUI_Factory.generate("3434".getBytes(StandardCharsets.UTF_8)).contains("HSP-"));
    assertEquals("HSP-14c879f3-f5d8-3d93-a09f-6090d77c2cc3",HSP_UUI_Factory.generate("3434".getBytes(StandardCharsets.UTF_8)));
  }
}
